+++ http/src/main/scala/com/twitter/finatra/http/response/StreamingResponse.scala
@@ -5,7 +5,7 @@ import com.twitter.finagle.http.{Response, Status}
 import com.twitter.inject.Logging
 import com.twitter.io.Reader.ReaderDiscarded
 import com.twitter.io.{Buf, Writer}
-import com.twitter.util.{Closable, Duration, Future}
+import com.twitter.util._
 
 import scala.util.control.NonFatal
 
@@ -19,12 +19,44 @@ object StreamingResponse {
   private val JsonArraySuffix = Some(Buf.Utf8("]"))
 
   /**
-   * Construct a [[StreamingResponse]] from an `AsyncStream`.
-   *
-   * A [[StreamingResponse]] is useful for streaming data back to the client in chunks: data will
-   * be rendered to the client as it resolves from the `AsyncStream` while also utilizing the
-   * back-pressure mechanisms provide by the underlying transport, preventing unnecessary resource
-   * consumption for properly constructed `AsyncStream`s.
+    * Construct a [[StreamingResponse]] from an `AsyncStream` and an `AsyncStream[T] => AsyncStream[(U, Buf)]`
+    *
+    * A [[StreamingResponse]] is useful for streaming data back to the client in chunks: data will
+    * be rendered to the client as it resolves from the `AsyncStream` while also utilizing the
+    * back-pressure mechanisms provide by the underlying transport, preventing unnecessary resource
+    * consumption for properly constructed `AsyncStream`s.
+    *
+    * @param streamTransformer Function which converts AsyncStream[T] to AsyncStream[(U, Buf)]
+    * @param status Status code of the generated response.
+    * @param headers Headers for the generated response.
+    * @param onDisconnect A hook to clean up resources upon disconnection, either normally or exceptionally.
+    * @param closeGracePeriod The grace period provided to close the Response.writer.
+    * @param asyncStream The data that will represent the body of the response.
+    * @tparam T The incoming type.
+    * @tparam U An auxillary type passed to onWrite which may be helpful when executing the onWrite callback.
+    *
+    */
+  def apply[T, U](
+    streamTransformer: AsyncStream[T] => AsyncStream[(U, Buf)],
+    status: Status,
+    headers: Map[String, String],
+    onWrite: (U, Buf) => Try[Unit] => Unit,
+    onDisconnect: () => Unit,
+    closeGracePeriod: Duration
+  )(asyncStream: => AsyncStream[T]): StreamingResponse[T, U] = {
+    new StreamingResponse[T, U](
+      streamTransformer,
+      status,
+      headers,
+      asyncStream,
+      onWrite,
+      onDisconnect,
+      closeGracePeriod
+    )
+  }
+
+  /**
+   * Construct a [[StreamingResponse]] from an `AsyncStream` and a `toBuf`
    *
    * @param toBuf Function for converting messages to a binary `Buf` representation.
    * @param status Status code of the generated response.
@@ -37,6 +69,7 @@ object StreamingResponse {
    * @param closeOnFinish A hook for cleaning up resources after completion of the rendering
    *                      process. Note that this will be called regardless of whether rendering
    *                      the body is successful.
+   * @param closeGracePeriod The grace period provided to the closeable in closeOnFinish
    * @param asyncStream The data that will represent the body of the response.
    */
   def apply[T](
@@ -46,18 +79,35 @@ object StreamingResponse {
     prefix: Option[Buf] = None,
     separator: Option[Buf] = None,
     suffix: Option[Buf] = None,
-    closeOnFinish: Closable = Closable.nop
-  )(asyncStream: => AsyncStream[T]) = {
-    new StreamingResponse[T](
+    closeOnFinish: Closable = Closable.nop,
+    closeGracePeriod: Duration = Duration.Zero
+  )(asyncStream: => AsyncStream[T]): StreamingResponse[T, Unit] = {
+
+    val chainedTransformers: AsyncStream[T] => AsyncStream[(Unit, Buf)] = {
+      (StreamingResponseUtils.toBufTransformer(toBuf) _)
+          .andThen(StreamingResponseUtils.separatorTransformer(separator) _)
+          .andThen(StreamingResponseUtils.prefixTransformer(prefix) _)
+          .andThen(StreamingResponseUtils.suffixTransformer(suffix) _)
+          .andThen(StreamingResponseUtils.tupleTransformer(()) _)
+    }
+
+    def onDisconnect(): Unit = {
+      closeOnFinish.close(closeGracePeriod)
+      ()
+    }
+
+    def onWrite(ignored: Unit, buf: Buf)(t: Try[Unit]): Unit = ()
+
+    StreamingResponse[T, Unit](
+      streamTransformer = chainedTransformers,
       status = status,
-      toBuf = toBuf,
       headers = headers,
-      prefix = prefix,
-      separator = separator,
-      suffix = suffix,
-      asyncStream = () => asyncStream,
-      closeOnFinish = closeOnFinish
-    )
+      onDisconnect = onDisconnect,
+      onWrite = onWrite,
+      closeGracePeriod = closeGracePeriod
+    ) {
+      asyncStream
+    }
   }
 
   def jsonArray[T](
@@ -66,35 +116,30 @@ object StreamingResponse {
     headers: Map[String, String] = Map(),
     closeOnFinish: Closable = Closable.nop,
     asyncStream: => AsyncStream[T]
-  ) = {
-
-    new StreamingResponse[T](
+  ): StreamingResponse[T, Unit] = {
+    StreamingResponse.apply(
       toBuf = toBuf,
       status = status,
       headers = headers,
       prefix = JsonArrayPrefix,
       separator = JsonArraySeparator,
       suffix = JsonArraySuffix,
-      asyncStream = () => asyncStream,
-      closeOnFinish = closeOnFinish
-    )
+      closeOnFinish = closeOnFinish) {
+        asyncStream
+    }
   }
 }
 
-/**
- * Representation of a streaming HTTP response based on the `AsyncStream` construct.
- */
-class StreamingResponse[T] private (
-  toBuf: T => Buf,
+class StreamingResponse[T, U] private (
+  streamTransformer: AsyncStream[T] => AsyncStream[(U, Buf)],
   status: Status,
   headers: Map[String, String],
-  prefix: Option[Buf],
-  separator: Option[Buf],
-  suffix: Option[Buf],
-  asyncStream: () => AsyncStream[T],
-  closeOnFinish: Closable,
-  closeGracePeriod: Duration = Duration.Zero
-) extends Logging {
+  asyncStream: => AsyncStream[T],
+  onWrite: (U, Buf) => Try[Unit] => Unit,
+  onDisconnect: () => Unit,
+  closeGracePeriod: Duration
+) extends Logging
+{
 
   def toFutureFinagleResponse: Future[Response] = {
     val response = Response()
@@ -104,11 +149,7 @@ class StreamingResponse[T] private (
     val writer = response.writer
 
     /* Orphan the future which writes to our response thread */
-    (for {
-      _ <- writePrefix(writer)
-      _ <- addSeparatorIfPresent(asyncStream().map(toBuf)).foreachF(writer.write)
-      result <- writeSuffix(writer)
-    } yield result).onSuccess { _ =>
+    write(writer).onSuccess { _ =>
       debug("Success writing to chunked response")
     }.onFailure {
       case e: ReaderDiscarded =>
@@ -117,33 +158,20 @@ class StreamingResponse[T] private (
         error("Unexpected failure writing to chunked response", e)
     }.ensure {
       debug("Closing chunked response")
-      Closable.all(response.writer, closeOnFinish).close(closeGracePeriod)
+      response.writer
+        .close(closeGracePeriod)
+        .ensure {
+          onDisconnect()
+        }
     }
 
     Future.value(response)
   }
 
-  private[this] def writePrefix(writer: Writer): Future[Unit] =
-    writeOption(prefix, writer)
-
-  private[this] def writeSuffix(writer: Writer): Future[Unit] =
-    writeOption(suffix, writer)
-
-  private[this] def writeOption(data: Option[Buf], writer: Writer): Future[Unit] = data match {
-    case Some(data) => writer.write(data)
-    case None => Future.Unit
-  }
-
-  private[this] def addSeparatorIfPresent(stream: => AsyncStream[Buf]): AsyncStream[Buf] =
-    separator match {
-      case Some(sep) => addSeparator(stream, sep)
-      case None => stream
+  private[this] def write(writer: Writer): Future[Unit] = {
+    streamTransformer(asyncStream).foreachF { case (item, buf) =>
+      writer.write(buf).respond(onWrite(item, buf))
     }
-
-  private[this] def addSeparator(stream: => AsyncStream[Buf], separator: Buf): AsyncStream[Buf] = {
-    stream.take(1) ++ (stream.drop(1).map { buf =>
-      separator.concat(buf)
-    })
   }
 
   private[this] def setHeaders(headersOpt: Map[String, String], response: Response) = {