+++ http/src/test/scala/com/twitter/finatra/http/tests/integration/tweetexample/main/controllers/TweetsController.scala
@@ -1,16 +1,22 @@
 package com.twitter.finatra.http.tests.integration.tweetexample.main.controllers
 
 import com.twitter.concurrent.AsyncStream
-import com.twitter.finagle.http.{Fields, Status, Request, Response}
+import com.twitter.finagle.http.{Fields, Request, Response, Status}
 import com.twitter.finatra.http.Controller
 import com.twitter.finatra.http.tests.integration.tweetexample.main.domain.Tweet
 import com.twitter.finatra.http.tests.integration.tweetexample.main.services.TweetsRepository
-import com.twitter.finatra.http.response.StreamingResponse
+import com.twitter.finatra.http.response.{StreamingResponse, StreamingResponseUtils}
 import com.twitter.io.Buf
-import com.twitter.util.Future
+import com.twitter.util.{Duration, Future, Try}
 import javax.inject.Inject
 
-class TweetsController @Inject()(tweetsRepository: TweetsRepository) extends Controller {
+import scala.collection.mutable
+
+class TweetsController @Inject()(
+  tweetsRepository: TweetsRepository,
+  onWriteLog: mutable.ArrayBuffer[String]
+)
+  extends Controller {
 
   get("/tweets/hello") { request: Request =>
     "hello world"
@@ -30,6 +36,49 @@ class TweetsController @Inject()(tweetsRepository: TweetsRepository) extends Con
 
   get("/tweets/streaming_custom_tobuf") { request: Request =>
     StreamingResponse(Buf.Utf8.apply) {
+      AsyncStream("A", "B", "C")
+    }
+  }
+
+  get("/tweets/streaming_with_transformer") { _: Request =>
+    def lowercaseTransformer(as: AsyncStream[String]) = as.map(_.toLowerCase)
+
+    val transformer = (lowercaseTransformer _)
+      .andThen(StreamingResponseUtils.toBufTransformer(Buf.Utf8.apply) _)
+      .andThen(StreamingResponseUtils.tupleTransformer(()) _)
+
+    def onWrite(ignored: Unit, buf: Buf)(t: Try[Unit]): Unit = ()
+
+    StreamingResponse(
+        transformer,
+        Status.Ok,
+        Map.empty,
+        onWrite,
+        () => (),
+        Duration.Zero
+      ) {
+      AsyncStream("A", "B", "C")
+    }
+  }
+
+  get("/tweets/streaming_with_onWrite") { _: Request =>
+
+    def serializeAndLowercase(as: AsyncStream[String]): AsyncStream[(String, Buf)] = {
+      as.map(str => (str.toLowerCase, Buf.Utf8(str)))
+    }
+
+    def onWrite(lowerCased: String, buf: Buf)(t: Try[Unit]): Unit = {
+      onWriteLog.append(lowerCased)
+    }
+
+    StreamingResponse(
+      serializeAndLowercase,
+      Status.Ok,
+      Map.empty,
+      onWrite,
+      () => (),
+      Duration.Zero
+    ) {
       AsyncStream("A", "B", "C")
     }
   }